+++ src/Clock/Clock.css
@@ -58,8 +58,8 @@
 	display: flex;
 	flex-direction: row;
 	align-items: center;
-	justify-content: space-evenly;
-	width: 95vw;
+	justify-content: space-between;
+	width: 60vw;
 }
 .start-button {
 	border: 2px solid white;
@@ -84,23 +84,6 @@
 	justify-content: center;
 	align-items: center;
 	cursor: pointer;
-}
-.clock-switch {
-	border: 2px solid white;
-	font-family: 'Major Mono Display', monospace;
-	width: auto;
-	padding-left: 2vw;
-	padding-right: 2vw;
-	height: 7vw;
-	font-size: 5vw;
-	color: white;
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-.clock-switch:hover {
-	background-color: rgb(243, 34, 73);
 }
 .start-button:hover {
 	background-color: rgb(34, 235, 158);
